% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zorks.R
\name{begin}
\alias{begin}
\title{Begin the game.}
\usage{
begin()
}
\description{
Type begin() to begin a game after loading a library, Q to quit. Otherwise
just use natural language. If you don't have an OpenAPI key, the package will
prompt you for one and save it in the `OPENAI_API_KEY` environment variable.
}
